added compile asset signals method
on occasions where the same asset appears multiple times in the holdings table, we need a way to condense signals to one value. This is not expected to happen much, but it should be added for robustness. the method compile_asset_signals is added. It first prioritises sell signals, then buy, then hold. Buy is prioritised second, because buy signals are essentially hold signals, if we already hold the asset